chore: move matetial type to common enum
