+++ src/api/material/dto/create-material.res.dto.ts
@@ -1,5 +1,5 @@
 import { ApiProperty } from '@nestjs/swagger'
-import { MaterialType } from '../entities/material.entity'
+import { MaterialType } from '@common/enums/material.enum'
 import { DateField, EnumField, StringField, UUIDField } from '@decorators/field.decorators'
 import { Expose } from 'class-transformer'
