+++ src/common/enums/material.enum.ts
@@ -1,0 +1,8 @@
+export enum MaterialType {
+  VIDEO = 'video',
+  PDF = 'pdf',
+  DOCUMENT = 'document',
+  IMAGE = 'image',
+  AUDIO = 'audio',
+  OTHER = 'other'
+}
